@@ -1,41 +1,54 @@
-name: flutter_memory_game
-description: A new Flutter project.
+name: endless_runner
+description: A Flame game template built with Flutter.
 
-# The following line prevents the package from being accidentally published to
-# pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: "none" # Remove this line if you wish to publish to pub.dev
+# Prevent accidental publishing to pub.dev.
+publish_to: 'none'
 
-version: 1.0.0+1
+version: 0.1.0
 
 environment:
-  sdk: ">=2.18.6 <3.0.0"
+  sdk: ^3.0.0
 
 dependencies:
-  confetti: ^0.7.0
-  cupertino_icons: ^1.0.2
-  flame: ^1.16.0
-  flame_audio: ^2.10.0
-  flip_card: ^0.6.0
+  flame: ^1.14.0
   flutter:
     sdk: flutter
-  gap: ^3.0.1
-  lottie: ^2.1.0
+
+  audioplayers: ^5.2.1
+  cupertino_icons: ^1.0.6
+  flame_audio: ^2.1.7
+  go_router: ^13.1.0
+  provider: ^6.1.1
+  shared_preferences: ^2.2.2
+  nes_ui: ^0.15.0
+  google_fonts: ^6.1.0
+  logging: ^1.2.0
 
 dev_dependencies:
-  flutter_lints: ^2.0.0
+  flame_test: ^1.15.3
+  flutter_lints: ^3.0.1
   flutter_test:
     sdk: flutter
+  flutter_launcher_icons: ^0.13.1
+  test: ^1.24.3
 
 flutter:
   uses-material-design: true
-  # To add assets to your application, add an assets section, like this:
+
   assets:
-    - assets/images/.
-    - assets/background_images/
-    - assets/audio/
-    - assets/brain_animation.json
-  #   - images/a_dot_ham.jpeg
+    - assets/images/
+    - assets/images/dash/
+    - assets/images/enemies/
+    - assets/images/scenery/
+    - assets/music/
+    - assets/sfx/
+
   fonts:
-    - family: Game     
+    - family: Press Start 2P
       fonts:
-        - asset: assets/fonts/Game.ttf
+        - asset: assets/fonts/Press_Start_2P/PressStart2P-Regular.ttf
+
+flutter_launcher_icons:
+  android: true
+  ios: true
+  image_path: "assets/icons/icon.png"
